refactor(api): Add Course routes to index structure
In this refactor will be necesary changes in the various files for not crash code
controler -> index.js
services -> index.js
provider -> index.js
courseRoutes.js
courseService.js
courseController.js
