+++ back/controllers/index.js
@@ -1,0 +1,3 @@
+const CourseController = require('./courseController');
+
+module.exports = { CourseController };
