+++ back/providers/index.js
@@ -1,0 +1,3 @@
+const courseProvider = require('./courseProvider');
+
+module.exports = { courseProvider };
