+++ back/services/index.js
@@ -1,0 +1,3 @@
+const courseService = require('./courseService');
+
+module.exports = { courseService };
